Add (unused) determine-aptible-identifiers function
This is some speculative work to simplify the burden of running this
script.

The information to map stack/env/app names to their ids is present in
the app, but a little tedious to get at. Now that the rest of the
problems are more or less solved, I wanted to tackle this one so that
you can just specify everything by name and not have to know the ids
or find them.