Checkpoint for no longer requiring manual aptible identifier mapping
Because determine-aptible-identifiers can use the app to determine
stack/env/app ids, I have integrated it into -main so that the edn
files don't have to be used to do it.

Before any security scans are done, the security-scan-tuples are
checked against the aptible-identifiers to ensure that all of them can
be determined. If not, the page will close and tell you which ones it
can't resolve, otherwise it will do the scans, like before.